added optimizer tests
