+++ cortex/built_ins/datasets/CelebA.py
@@ -13,12 +13,8 @@ from .utils import build_transforms
 class CelebAPlugin(DatasetPlugin):
     sources = ['CelebA']
 
-    def handle(
-            self,
-            source,
-            copy_to_local=False,
-            normalize=True,
-            **transform_args):
+    def handle(self, source, copy_to_local=False, normalize=True,
+               **transform_args):
         Dataset = self.make_indexing(CelebA)
         data_path = self.get_path(source)
